requirements.txt
streamlit
freesound-api
google-generativeai==0.3.2